tests/wpt/tests/shared-storage/cross-origin-create-worklet-credentials-omit.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestor_key = token();
  const crossOrigin = 'https://{{domains[www]}}:{{ports[https][0]}}';
  const set_cookie_url = crossOrigin + `/cookies/resources/set-cookie.py` +
                         `?name=key0` +
                         `&path=/shared-storage/`;
  const helper_url = crossOrigin +
                     `/shared-storage/resources/credentials-test-helper.py` +
                     `?access_control_allow_origin_header=${window.origin}` +
                     `&shared_storage_cross_origin_worklet_allowed_header=?1` +
                     `&token=${ancestor_key}`;

  await fetch(set_cookie_url, { mode: 'no-cors', credentials: 'include' });

  const worklet = await sharedStorage.createWorklet(
    helper_url + `&action=store-cookie`,
    { credentials: "omit" });

  const request_cookie_fetch_response =
    await fetch(helper_url + `&action=get-cookie`);

  const request_cookie_text = await request_cookie_fetch_response.text();

  assert_equals(request_cookie_text, "NO_COOKIE_HEADER");
}, 'createWorklet() with cross-origin module script and credentials "omit"');

</script>
</body>
